% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{add_project_id}
\alias{add_project_id}
\title{Add Project ID}
\usage{
add_project_id(data_set, project_id)
}
\arguments{
\item{data_set}{The data set being processed}

\item{project_id}{The ID being assigned to the project (see function description)}
}
\value{
Returns the dataframe with a column for project ID
}
\description{
Adds a Project ID to the RHoMIS data. The ID comes in the form AA_BBB_YYYY.
Where AA is the two letter ISO country-code (https://www.iban.com/country-codes)
Where BBB is a three letter code assigned to the project being processed
Where YYYY is the year the survey was conducted
}
\examples{
df <- add_project_id(df, "KE_TST_2016")
}
