% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{age_loop_calculation}
\alias{age_loop_calculation}
\title{Age loop conversion}
\usage{
age_loop_calculation(data_set)
}
\arguments{
\item{data_set}{the data_set that you want to convert}
}
\value{
Returns a data dataframe with household members placed into categories e.g. "children aged 4 to 10"
}
\description{
Age loop conversion
}
\examples{

df <- age_loop_calculation(data_set)
}
