% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{read_raw_data}
\alias{read_raw_data}
\title{RHoMIS Data Cleaning Functions}
\usage{
read_raw_data(path)
}
\arguments{
\item{path}{The file-path where the RHoMIS raw data is located}
}
\value{
Data-frame of rhomis dataset.
}
\description{
A series of functions that can be used to clean RHoMIS data.
read raw rhomis data
}
\details{
A function for loading raw RHoMIS data.
}
\examples{
df <- read_raw_data('C:/User/me/my_data.csv')

}
