% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{which_hdds_type_used}
\alias{which_hdds_type_used}
\title{Which HDDS}
\usage{
which_hdds_type_used(data_set)
}
\arguments{
\item{data_set}{The dataset where we need to identify which hdds type used}
}
\value{
Returns '10' if 10 groups were used. Returns '14' if 14 groups were used.
}
\description{
RHoMIS has used two indicators for HDDS. One made up of 10 food groups, and one made up of 14 food-groups
}
\examples{
hdds_type <- which_hdds_type_used(data_set)
}
